Remove code duplication in the functional interface
